api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
pydantic>=2.6
pydantic-settings>=2.2
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
httpx>=0.27
opentelemetry-api>=1.24
opentelemetry-sdk>=1.24
opentelemetry-instrumentation-fastapi>=0.45b0
prometheus-client>=0.20
